--- packages/@headlessmedia-react/playground/src/App.module.css
@@ -1,15 +1,0 @@
-.wrapper {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.buttonWrapper {
-  display: flex;
-  width: 600px;
-  align-items: center;
-  justify-content: space-between;
-}
