--- packages/@headlessmedia-react/src/constants.ts
@@ -1,8 +1,0 @@
-export const DEFAULT_AUTO_BITRATE_INDEX = -1
-
-export enum MediaStatus {
-  LOADING = 'LOADING',
-  CAN_PLAY = 'CAN_PLAY',
-  RECOVERING = 'RECOVERING',
-  ERROR = 'ERROR',
-}
