+++ packages/@headlessmedia-vue/playground/src/App.vue
@@ -1,12 +1,6 @@
 <template>
   <Control />
-  <video
-    width="800"
-    height="400"
-    controls
-    src="http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4"
-    v-on="mediaEventHanlders"
-  />
+  <Video />
 </template>
 
 <script lang="ts">
@@ -14,19 +8,11 @@ import { defineComponent } from 'vue'
 import { useMedia } from '@headlessmedia/vue'
 
 import Control from './components/Control.vue'
+import Video from './components/Video.vue'
 
 import { mediaId } from './const'
 
 export default defineComponent({
-  components: {
-    Control,
-  },
-  name: 'App',
-  setup() {
-    const { getMediaProps } = useMedia({ id: mediaId })
-    const mediaEventHanlders = getMediaProps()
-
-    return { mediaEventHanlders }
-  },
+  components: { Control, Video },
 })
 </script>
